SitesInfected, SitesDisturbed, SitesDamaged moved to Summary log Added LethalTemp column to Summary log, which represents the proportion of active cells where the pathogen is absent due to low temp Event log includes species biomass removed per timestep New map output LethalTemp shows the low temp where it caused the elimination of the pathogen. Removed dependency on Climate Library

+++ src/IntPIxel.cs
@@ -4,7 +4,7 @@ namespace Landis.Extension.RootRot
 {
     public class IntPixel : Pixel
     {
-        public Band<double> MapCode = "The numeric code for each raster cell";
+        public Band<int> MapCode = "The numeric code for each raster cell";
 
         public IntPixel()
         {
